Updated Button Library
I have completely removed all of the existing buttons from the button library because now we have decided to go with only three generic buttons

- WOITextButton
- WOIIconButtton
- WOIParallelogramButton

Now buttons are aligned with the requirements and their naming convention is also consistent with initials WOI (We Over I).
