--- src/docs/loading-button.mdx
@@ -1,13 +1,0 @@
-# Loading Button Component
-
-The loading button component is a reusable React component for creating interactive buttons in your web application.
-
-## Props
-
-The Loading Button component accepts the following props:
-
-- `width` (number): The width of the button, specified as a numerical value (e.g., pixels) or as a percentage of its container's width.
-- `height` (number): The height of the button, specified as a numerical value (e.g., pixels) or as a percentage of its container's height.
-- `backgroundColor` (string): TThe background color of the button. You can specify this using a CSS color value, such as a Hex code (#RRGGBB), an RGB color (rgb(R, G, B)), or a named color.
-- `loaderColor` (string): The text color of the button when the mouse pointer hovers over it. This allows you to change the text color on hover.
-- `borderRadius` (number): The borderRadius CSS property rounds the corners of an element's outer border edge. You can set a single radius to make circular corners, or two radii to make elliptical corners.
