--- src/docs/toggle-button.mdx
@@ -1,11 +1,0 @@
-# Toggle Button Component
-
-The toggle button component is a reusable React component for creating interactive buttons in your web application.
-
-## Props
-
-The Toggle Button component accepts the following props:
-
-- `isToggled` (boolean): The state of the button.
-- `trueBackgroundColor` (string): The true state color of the button. You can specify this using a CSS color value, such as a Hex code (#RRGGBB), an RGB color (rgb(R, G, B)), or a named color.
-- `falseBackgroundColor` (string): The false state color of the button. You can specify this using a CSS color value, such as a Hex code (#RRGGBB), an RGB color (rgb(R, G, B)), or a named color.
