--- src/stories/icon-button.stories.ts
@@ -1,29 +1,0 @@
-
-import IconButton from '../components/IconButton';
-
-export default {
-    title: 'Icon Button',
-    component: IconButton,
-    tags: ['autodocs'],
-};
-
-export const Default = {
-    args: {
-        text: 'Icon Button',
-        textColor: '#ffffff',
-        hoverTextColor: '#e64178',
-        fontSize: 16,
-        fontWeight: 400,
-        width: 200,
-        height: 50,
-        backgroundColor: '#e64178',
-        hoverBackgroundColor: '#ffffff',
-        textTransform: 'capitalize',
-        borderRadius: 50,
-        openLink: 'https://www.wikipedia.org/',
-        openLinkInNewTab: true,
-        prefixIcon: 'https://cdn-icons-png.flaticon.com/512/271/271220.png',
-        suffixIcon: 'https://cdn-icons-png.flaticon.com/512/32/32213.png',
-        isDisabled: false,
-    },
-};
