--- src/stories/loading-button.stories.ts
@@ -1,18 +1,0 @@
-
-import LoadingButton from '../components/LoadingButton';
-
-export default {
-    title: 'Loading Button',
-    component: LoadingButton,
-    tags: ['autodocs'],
-};
-
-export const Default = {
-    args: {
-        width: 220,
-        height: 50,
-        backgroundColor: '#e64178',
-        loaderColor: '#6d5252',
-        borderRadius: 8
-    },
-};
