--- src/stories/parallelogram-button.stories.ts
@@ -1,30 +1,0 @@
-
-import ParallelogramButton from '../components/ParallelogramButton';
-
-export default {
-    title: 'Parallelogram Button',
-    component: ParallelogramButton,
-    tags: ['autodocs'],
-};
-
-export const Default = {
-    args: {
-        text: 'Parallelogram Button',
-        textColor: '#ffffff',
-        hoverTextColor: '#e64178',
-        fontSize: 16,
-        fontWeight: 400,
-        width: 300,
-        height: 48,
-        backgroundColor: '#e64178',
-        hoverBackgroundColor: '#ffffff',
-        textTransform: 'capitalize',
-        openLink: 'https://www.wikipedia.org/',
-        openLinkInNewTab: false,
-        prefixIcon: 'https://cdn-icons-png.flaticon.com/512/271/271220.png',
-        suffixIcon: 'https://cdn-icons-png.flaticon.com/512/32/32213.png',
-        isDisabled: false,
-        skewType: 'left',
-        skew: 20
-    },
-};
