--- src/stories/rectangle-button.stories.ts
@@ -1,26 +1,0 @@
-
-import RectangleButton from "../components/RectangleButton";
-
-export default {
-    title: 'Rectangle Button',
-    component: RectangleButton,
-    tags: ['autodocs'],
-};
-
-export const Default = {
-    args: {
-        text: 'Rectangle Button',
-        textColor: '#ffffff',
-        hoverTextColor: '#e64178',
-        fontSize: 16,
-        fontWeight: 400,
-        width: 200,
-        height: 48,
-        backgroundColor: '#e64178',
-        hoverBackgroundColor: '#ffffff',
-        textTransform: 'capitalize',
-        openLink: 'https://www.wikipedia.org/',
-        openLinkInNewTab: false,
-        isDisabled: false
-    },
-};
