--- src/stories/toggle-button.stories.ts
@@ -1,16 +1,0 @@
-
-import ToggleButton from '../components/ToggleButton';
-
-export default {
-    title: 'Toggle Button',
-    component: ToggleButton,
-    tags: ['autodocs'],
-};
-
-export const Default = {
-    args: {
-        isToggled: false,
-        trueBackgroundColor: '#40e1e6',
-        falseBackgroundColor: '#cccccc',
-    },
-};
